Added fade in feature for navbar
when the sign in page is clicked. It will load the navbar. In the future if the user is already logged in there will be not sign in page and the navbar will fade in automatically. If they're not logged in they will have to go through the account system and then will it load in.